allow serivce names to be used for config endpoint
