+++ src/chia-dotnet/ServicesNames.cs
@@ -1,0 +1,17 @@
+
+namespace chia.dotnet
+{
+    /// <summary>
+    /// The names of chia services. These are used as <see cref="Message.Destination"/> values
+    /// </summary>
+    public struct ServiceNames
+    {
+        public const string FullNode = "chia_full_node";
+        public const string Wallet = "chia_wallet";
+        public const string Farmer = "chia_farmer";
+        public const string Harvester = "chia_harvester";
+        public const string Simulator = "chia_full_node_simulator";
+        public const string Plotter = "chia plots create";
+        public const string Daemon = "daemon";
+    }
+}
